ChiEFTint

Functions needed to generate nucleon-nucleon (NN) potential from Chiral EFT.

The parameters needed for chiEFTint are specified in optional_parameters.jl. If you want to change Low-energy constatants, edit the LECs.jl or use dictionary for LECs e.g., in main_chiEFTint.jl.

NuclearToolkit.Gauss_Legendre — Method

Gauss_Legendre(xmin,xmax,n;eps=3.e-16)

to calculate mesh points and weights for Gauss-Legendre quadrature

NuclearToolkit.QL — Method

QL(z,J::Int64,ts,ws)

To calculate Legendre functions of second kind by Gauss-Legendre quadrature

NuclearToolkit.RKstep — Method

RKstep(T,Ho,eta,R,faceta,fRK,Ht)

wrapper function to calc. a Runge-Kutta (RK) step

NuclearToolkit.Rnl_all_ab — Method

Rnl_all_ab(lmax,br,n_mesh,xr_fm)

Returns array for radiul functions (prop to generalized Laguerre polynomials) HO w.f. in momentum space. Rnlk(l,n,k)=sqrt(br) * R(n,L,Z) Z with Z=brk (k=momentum in fm^-1)

NuclearToolkit.SRG — Method

SRG(xr,wr,V12mom,dict_numst,to)

main function for Similarity Renormalization Group (SRG) transformation of NN interaction in CM-rel momentum space.

NuclearToolkit.calc_Vmom! — Method

calc_Vmom!(pnrank,V12mom,tdict,xr,LEC,LEC2,l,lp,S,J,pfunc,n_reg,to;is_3nf=false)

calc. nn-potential for momentum mesh points

NuclearToolkit.commutator — Method

commutator(A,B,R,fac)

wrapper function to overwrite R by fac*(AB-BA)

NuclearToolkit.hw_formula — Method

hw_formula(A,fnum)

empirical formula for harmonis oscillator parameter hw by mass number A fnum=2: for sd-shell, Ref. J. Blomqvist and A. Molinari, Nucl. Phys. A106, 545 (1968).

NuclearToolkit.make_chiEFTint — Method

make_chiEFTint(;is_plot=false,BO=false,optHFMBPT=false,itnum=20,writesnt=true,nucs=[])

main function in chiEFTint. This generates NN-potential in momentum space and then transforms it in HO basis. This function is exported and can be simply called make_chiEFTint() in the run script.

Optional arguments: Note that these are mainly for the author's purposes, so you do not specify these.

optHFMBPT::Bool to optimize LECs using BO&HFMBPT

itnum::Int=20 number of iteration for BO&HFMBPT

is_plot::Bool, to visualize optimization process of LECs

writesnt::Bool, to write out interaction file in snt (KSHELL) format. julia writesnt = false case can be usefull when you repeat large number of calculations for different LECs.

NuclearToolkit.read_LECs! — Method

read_LECs!(LECs,idxLECs,dLECs;initialize=false,inpf="src/chiEFTint/LECs.jl")

read LECs from "src/chiEFTint/LECs.jl" (default) and store them as LECs(Vector{Float}), idxLECs(Vector{Int}), and dLECs(Dict{str,Float}).

NuclearToolkit.srg_tr — Method

srg_tr(Ho,T,Ht,Hs,eta,R,sSRG,face,ds,numit,to; r_err=1.e-8,a_err=1.e-8,tol=1.e-6)

to carry out SRG transformation

NuclearToolkit.OPEP — Method

OPEP(ts,ws,xr,V12mom,dict_numst,to,lsjs,llpSJ_s,tllsj,opfs;pigamma=true)

calc. One-pion exchange potential

NuclearToolkit.single_tpe — Method

single_tpe(nd_mpi,nd_mpi2,nd_mpi4,nd_mpi6,nd_mpi8,Fpi2,Fpi4,Fpi6,c1,c2,c3,c4,r_d12,r_d3,r_d5,r_d145,J,pnrank,ts,ws,xdwn,ydwn,xdwn2,ydwn2,k2,pjs,gis,opfs,fc,f_idx,tVs,lsj,tllsj,tdict,V12mom,V_i,V_j,to)

TPE contribution in a given momentum mesh point

NuclearToolkit.tpe — Method

tpe(LECs,ts,ws,xr,V12mom,dict_numst,to,llpSJ_s,lsjs,tllsj,opfs)

calc. two-pion exchange terms up to N3LO

NuclearToolkit.CGm0 — Method

CGm0(l1,l2,l)

CG coefficients for special case (l1,m1=0,l2,m2=0|l,m=0)

NuclearToolkit.HObracket — Method

HObracket(n1,l1,n2,l2,n,l,N,L,Lam;d=1)

To calc. harmonic oscillator brackets, <EL,el:Lam|e1l1,e2l2:Lam>d. The phase (-1)^(N+n+n1+n2) is needed to reproduce Tab.1 of [1]. This function has large overlap to other functions (gmosh, etc.), so it will be merged in near future...

Reference:

[1] B.Buck& A.C.merchant, Nucl.Phys. A600 (1996) 387-402

[2] G.P.Kamuntavicius et al., Nucl.Phys. A695 (2001) 191-201

NuclearToolkit.PreCalc6j — Method

PreCalc6j(emax)

preallocate 6j used for HF

\[\begin{Bmatrix} j_a & j_b & J \\ j_c & j_d & J' \end{Bmatrix}\]

$ja,jb,jc,jd$ are half-integer extended for kinetic_tb

\[\begin{Bmatrix} j_1/2& j_2/2& 1 \\ l_2& l_1& 1/2 \end{Bmatrix}\]

are needed to get dict6j[J][key] with key = [ja,jb,jd,jc,Jp]. Using array as key is in general slow, so the key is transformed into integer in the IMSRG part (But, this sometimes reduces readability...)

NuclearToolkit.PreCalcHOB — Method

PreCalcHOB(chiEFTobj,to)

see also "struct HarmonicOscillatorBrackets" in hartreefock.jl/def_struct.jl

In the early version, dict9j is defined as dict9j = [ [ Dict{Vector{Int64},Float64}() for S=0:1 ] for J=0:Jmax] with key~[la,ja,lb,jb,L] Dict using array as key is slow, so this was fixed to... dict9j => nested array J=>S=>L=>ja=>la=>jb=>lb Note that div(j,2)+1 will be used as idx for ja&jb.

The same can be said for HOBs HOBs => nested array N=>n=>Lam=>lam=>L=>na=>nb=>la (lb is automatically determined)

NuclearToolkit.TMtrans — Method

TMtrans(dLECs,xr,wr,xrP,wrP,Rnl,RNL,nTBME,infos,izs_ab,Numpn,V12ab,arr_numst,dict6j,d6j_nabla,X9,U6,to;calc_relcm=false,writesnt=true)

Function to carry out Talmi-Mochinsky transformation

NuclearToolkit.kinetic_ob — Method

kinetic_ob(nlj1, nlj2)

calc. kinetic one-body contribution

NuclearToolkit.kinetic_tb — Method

kinetic_tb(nljtz1, nljtz2, nljtz3, nljtz4,J, dict6j,d6j_nabla,key6j)

calc. kinetic two-body contribution using preallocated 6j Dict

NuclearToolkit.prep_wsyms — Method

prep_wsyms()

cg1s = (1,0,l,0|l',0), cg2s = (2,0,l,0|l',0)

NuclearToolkit.prepareX9U6 — Method

prepareX9U6(Nrmax;to=nothing)

return 6j/9j dict: for 6j => jj->S->lam->lc->lr, for 9j => S->J->key= [la,nja,lb,njb,lam]

NuclearToolkit.readsnt — Method

readsnt(sntf,Anum;eachA=false,pnfac=1.0)

to read sntfile. This is slightly different from readsnt() in ShellModel.jl

NuclearToolkit.s_wigner9j — Method

s_wigner9j(j1,j3,j4,j6,j7,j8,j9)

to calc. wigner9j for specific cases with j2=j5=1/2

NuclearToolkit.wigner9j — Method

wigner9j(j1,j2,j3,j4,j5,j6,j7,j8,j9)

calculate Wigner 9j symbols, all j should be given as integer (0,1,...) or halfinteger (3/2, 3//2,...)

NuclearToolkit.chiEFTparams — Type

chiEFTparams

mutable struct to specify parameters for chiEFTint

Fields

NuclearToolkit.init_chiEFTparams — Method

init_chiEFTparams(;fn_params="optional_parameters.jl")

constructor of chiEFTparams, see chiEFTparams mutable struct for more details.

NuclearToolkit.momplot — Method

momplot(xr,V12mom,tdict,pnrank,llpSJ_s;ctext="",fpath="")

plot nn potential in partial wave over relative momentum space

NuclearToolkit.print_vec — Method

print_vec(s,v;ine=false)

function to make println(stdout) more readable. This is usuful for debug.

NuclearToolkit.pw_plt — Method

pw_plt(tx,xr,z,zds,pnrank;fdat=[])

fdat: normally no need to specify. optional array to specify text file for cross check

NuclearToolkit.read_chiEFT_parameter! — Method

read_chiEFT_parameter!(fn,params)

Function to overwrite params from the parameter file fn.

NuclearToolkit.write_tbme — Method

write_tbme(chiEFTobj,io,ndim,izs,Jtot,vv,nljsnt,nljdict,tkeys,dict6j,d6j_nabla,key6j;ofst=0)

write tbme in myg/snt(snt.bin) format

NuclearToolkit.write_vmom — Method

write_vmom(xr,V12mom,tdict,pnrank,llpSJ_s;label="")

NuclearToolkit.anti_op_isospin — Method

anti_op_isospin(dWS,n12,l12,s12,j12,t12,n3,l3,j3,n45,l45,s45,j45,t45,n6,l6,j6,jtot,ttot,to)

Function to calc. matrix element of antisymmetrizer. Detailed derivation can be found in e.g., Eq.(3.119) of Master Thesis by Joachim Langhammer (2010), TUDarmstadt.

NuclearToolkit.overwrite_key6! — Method

overwrite_key6!(s12,S,s45,key)

overwrite key for 6j symbols

NuclearToolkit.overwrite_key9! — Method

overwrite_key9!(l12,s12,j12,l3,j3,Lam,stot,jtot,key)

overwrite key for 9j symbols

NuclearToolkit.prep_dWS3n — Method

prep_dWS3n(e3max,N3max)

to prepare Wigner symbols (CG/6j/9j) for three body force

NuclearToolkit.calc_GPllh — Method

-0.5 y^TKtt^{-1}y - logdet(Kw)

NuclearToolkit.XF — Method

Refs.: