IMSRG

Functions for IM-SRG calculations

imsrg_util.jl: contains main and util functions

commutator.jl: functions to calculate commutators and BCH transform to carry out IMSRG flow with Magnus expansion

valencespace.jl: functions for Valence-space IM-SRG (VS-IMSRG) calculations to derive shell-model effective interactions/operators

NuclearToolkit.BCH_Product — Method

BCH_Product(X,Y,Z,tmpOp,Nested,ncomm,norms,Chan1b,Chan2bD,HFobj,dictMono,dict6j,PandyaObj,to;tol=1.e-4)

returns $Z$ to satisfy: $e^Z = e^X e^Y$

$Z$ is calculated with Baker–Campbell–Hausdorff (BCH) formula:

$Z = X + Y + 1/2[X, Y] + 1/12 [X,[X,Y]] + 1/12 [Y,[Y,X]] -1/24 [Y,[X,[X,Y]]] -1/720 [Y,[Y,[Y,[Y,X]]]] -1/720 [X,[X,[X,[X,Y]]]] +...$

For IMSRG flow of $H(s)$, $X=\eta(s)*ds$, $Y=\Omega(s)$, and $Z=\Omega(s+ds)$

NuclearToolkit.BCH_Transform — Method

BCH_Transform(Omega,O0,Hs,tOp,Nested,ncomm,norms,Chan1b,Chan2bD,HFobj,dictMono,dict6j,PandyaObj,to;tol=1.e-9,maxit=50,verbose=false)

Update $B$ (assumed to be $H$ or $O$) via $e^ABe^{-A} =B+[A,B]+1/2![A,[A,B]]+1/3![A,[A,[A,B]]]+...$

It needs 3 copy to store (k-1)-th adj./k-th adj./ret.

Note that the ret and tOp are also overwritten. BCH_Transform(nOmega,H0,IMSRGObj,tmpOp,norms,Chan2b) e.g., Omega: $\Omega(s+ds)$, H0: $H(s=0)$ or $O(s=0)$, and ret: $H(s+ds)$

NuclearToolkit.OpCommutator! — Method

OpCommutator!(X::Op,Y::Op,ret::Op,HFobj,Chan1b,Chan2bD,dictMono,dict6j,PandyaObj,to) where{Op <: Operator}

overwrite ret operator by $[X,Y]$

NuclearToolkit.calcZbar! — Method

calcZbar!(Xbar,Ybar,PhaseMat,PhaseMatY,tmpMat,hy,nph_kets,nKets_cc,Zlefthalf,Zrighthalf,hz)

`Xbar`: (`nKets_cc`, 2*`nph_kets`)

Ybar: (2nph_kets,nKets_cc) PhaseMatY: (nph_kets,nKets_cc) Zbar: (nKets_cc, 2nKets_cc)

                | nph_kets,nKets_cc

(nKetscc,nKetscc) |–––––– | nphkets,nKetscc

NuclearToolkit.comm111ss! — Method

comm111ss!(X,Y,ret;inifac=1.0)

$[X_1,Y_1]$

NuclearToolkit.comm121ss! — Method

comm121ss!(X,Y,ret,HFobj,Chan1b,Chan2b,dictMono,PandyaObj)

returns $[X_1,Y_2] - [Y_1,X_2]$, whose elements are given as

$[X_1,Y_2]_{ij} = \frac{1}{2j_i+1}\sum_{ab} (n_a \bar{n}_b) \sum_{J} (2J+1) (X_{ab} Y^J_{biaj} - X_{ba} Y^J_{aibj})$

NuclearToolkit.comm220ss! — Method

comm220ss!(X,Y,Z::zerobody,HFobj,Chan2b) where {zerobody<:Vector{Float64}}

$[X_2,Y_2]_0 = 2 \sum_{J}(2J+1) \mathrm{Tr}(X_{hh'pp'} Y_{pp'hh'})$

NuclearToolkit.comm221ss! — Method

comm221ss!(X,Y,ret,HFobj,Chan1b,Chan2bD,PandyaObj)

returns $[X_2,Y_2]_1 - [Y_2,X_2]_1$, whose elements are given as

$[X_2,Y_2]_{ij} = 1/(2[j_i]) \sum_{abc}\sum_{J}[J](n'_an'_bn_c-n_an_bn'_c)(X_{2,ciab}Y_{2,abcj}-Y_{2,ciab}X_{2,abcj})$

Note that our code gives diffrent results from the IMSRG code by S.R.Stroberg here. If you replace sqfac_ab^2 in this function by 2.0, the results agree with each other.

NuclearToolkit.single_121 — Method

single_121(a,b,i,j,o1b,o2bs,sps,key,targetDict;verbose=false)

calc 121part $[X_1,Y_2]-[Y_1,X_2]$ for given i,j and a,b.

$\sum_{J} [J]^2 (o1_{ab}o2_{biaj} - o1_{ba}o2_{aibj})$

NuclearToolkit.DoPandyaTransformation_SingleChannel — Function

function to carry out Pandya transformation:

NuclearToolkit.GatherOmega — Method

GatherOmega(Omega,nOmega,gatherer,tmpOp,Nested,H0,Hs,ncomm,norms,Chan1b,Chan2bD,HFobj,IMSRGobj,dictMono,dict6j,PandyaObj,maxnormOmega,to)

This may not be used now.

NuclearToolkit.Get2bDenominator — Method

Get2bDenominator(ch,pnrank,a,b,i,j,na,nb,ni,nj,f,Delta,dictMono,key;verbose=false)

$f_{aa} +f_{bb} −f_{ii} −f_{jj} +G_{abij} +\Delta$

with $G_{abij} = \Gamma_{abab} + \Gamma_{ijij} - (\Gamma_{aiai} + \Gamma_{bjbj} + [a \leftrightarrow b])$

NuclearToolkit.Gethhph — Method

Gethhph(kets,sps)

get idxs for hh/ph kets

NuclearToolkit.adhoc_rewrite6jdict — Function

adhoc_rewrite6jdict(emax,dict6j,ofst_unit=1000)

adhoc function to replace dict. for 6j-symbols:{ji,jj,J,jk,jl,J'} key = [ji,jj,jk,jl,J'] -> newkey::Int64 newdict is Vector for total J

NuclearToolkit.calc_Eta_atan! — Method

calc_Eta_atan!(HFobj,IMSRGobj,Chan2b,dictMono,norms)

calc. $\eta(s)$ with atan generator

NuclearToolkit.check_order_Mkey — Method

check_order_Mkey(key,pnrank)

reorder key to be key[1] > key[2]

NuclearToolkit.flow_Operators — Method

flow_Operators(binfo,HFobj,IMSRGobj,PandyaObj,Chan1b,Chan2bD,dict_9j,HOBs,dictMono,dict6j,Operators,to)

consistent IMSRG flow of scaler operators (Rp2) using written Omega

NuclearToolkit.generator_atan — Method

generatoratan a: particle index i: hole index Gam: D'NO2B iket:: index for |ai> to get Gam{aiai}=Gam[iket,iket] Delta: parameter for energy shift like EKK

NuclearToolkit.getNorm — Method

getNorm(O,p_sps,n_sps,Chan2b)

returns sqrt(norm1b^2 + norm2b^2)

NuclearToolkit.getNorm1b — Function

getNorm1b(Mat1b,p_sps,n_sps,verbose=false)

returns 1bnorm of the given Operator

NuclearToolkit.getNorm2b — Function

getNorm2b(Mat2b,Chan2b,verbose=false)

returns 2bnorm of the given Operator

NuclearToolkit.imsrg_main — Method

imsrg_main(binfo,Chan1b,Chan2bD,HFobj,dictMono,d9j,HOBs,dict6j,valencespace,Operators,to; core_generator_type="atan",valence_generator_type="shell-model-atan",denominatorDelta=0.0)

Arguments

binfo::basedat struct basedat(nuc::nuclei,sntf::String,hw::Float,emax::Int)

Chan1b::chan1b struct for one-body stuffs

Chan2bD::chan2bD struct for two-body stuffs (e.g., dict to get idx from JPT)

HFobj::HamiltonianNormalOrdered struct HNO, which includes info. of HF solution (HF energy, occupation, f,Gamma,...)

dictMono::Dict dictionary to get Vmonopole

d9j preallocated dictionaries for wigner 9j symbols, which are needed to calculate Rp2

HOBs struct HarmonicOscillatorBrackets

dict6j preallocated dictionaries for wigner6j symbols (needed in e.g., Pandya transformation)

valencespace to specify valence space

Operators::Vector{String} non-Hamiltonian operators

to TimerOutput object to measure runtime&memory allocations

Optional Arguments

core_generator_type only the "atan" is implemented

valence_generator_type only the "shell-model-atan" is implemented

denominatorDelta::Float denominator Delta, which is needed for multi-major shell decoupling

NuclearToolkit.init_IMSRGobject — Method

init_IMSRGobject(HFobj;smax=500.0,dsmax=0.5,maxnormOmega=0.25,tol=1.e-6,eta_criterion=1.e-6,denominatorDelta=0.0)

Constructor for IMSRGobject

H0::Operator for starting point of BCH product

H::Operator Hamiltonian $H(s)$

s::Vector{Float} current $s$ and $ds$

smax::Float maximum $s$

dsmax::Float maximum $ds$

maxnormOmega::Float maximum ||Omega||

eta::Operator generator of IMSRG flow (antihermite Operator)

Omega::Operator generator of IMSRG flow (antihermite Operator)

eta_criterion::Float ||eta|| to check convergence

denominatorDelta::Float64 parameter for multi-major shell decoupling

n_written_omega::Int # of written Omega by splitting to solve IMSRGflow

Ncomm::Vector{Int} # of commutator evaluated during IMSRG flow

NuclearToolkit.init_dictMonopole! — Method

init_dictMonopole!(dictMonopole,Chan2b)

initialize dictMonopole

NuclearToolkit.make_PandyaKets — Method

make_PandyaKets(emax,HFobj)

To prepare "kets" for Pandya transformation. For ordinary two-body channels, kets like |i,j=i;J=odd> where i={n,l,j,tz} are hindered, but necessary for Pandya transformation.

NuclearToolkit.prep_PandyaLookup — Method

numbersPandya:[ch,nKetcc,nhh,nph] for ich (channel index of Chan2b_Pandya)

NuclearToolkit.print_flowstatus — Method

print_flowstatus(istep,s,ncomm,norms,IMSRGobj)

print flowstatus s,E0,1b&2b norm for Omega, 1b&2b norm for Eta, Ncomm, nwritten

NuclearToolkit.read_imsrg_parameter! — Method

read_imsrg_parameter!(fn,IMSRGobj)

Function to overwrite IMSRGobj from the parameter file fn.

NuclearToolkit.read_omega_bin! — Function

read_omega_bin!(nw,Op,verbose=false)

read written Omega file and update Op::Operator

NuclearToolkit.set_dictMonopole! — Method

set_dictMonopole!(dictMonopole,HFobj,H)

To update dictMonopole pp/pn/nn under H(s=0)/IMSRG H(s)

NuclearToolkit.set_sps_to_core! — Method

set_sps_to_core!(binfo,HFobj)

modify p_sps.occ, n_sps.occ by the "core" nucleus

NuclearToolkit.set_sps_to_modelspace! — Method

set_sps_to_modelspace!(binfo,HFobj)

modify occupation by specified model space

NuclearToolkit.update_core_in_sps! — Method

update_core_in_sps!(binfo,HFobj)

Function to specify hole/core for sps. This will will be used for target normal ordering

NuclearToolkit.write_omega_bin — Method

to write binary file for Operator flow using split/huntergatherer method

NuclearToolkit.calc_Eta_smatan! — Method

calc_Eta_smatan!(HFobj,IMSRGobj,Chan2b,dictMono,norms)

$\eta(s)$ with shell-model atan generator to decouple valence space

NuclearToolkit.check_major_valencespace — Method

check_major_valencespace(str::String,HFobj,v)

Function to check valence space and overwrite v and q fields of SingleParticleState The valencespace is specified by argument str (e.g. "p-shell")

NuclearToolkit.check_str_valencespace — Method

check_str_valencespace(valencespace::Vector{Vector{Int64}},HFobj,v)

check valence space and overwrtie SingleParticleState.v/q

specified by or Vector{Int} (e.g., [[0,1,1,-1],[0,1,3,-1], [0,1,1,1],[0,1,3,1]])

NuclearToolkit.check_valence_space — Method

check_valence_space(HFobj,valencespace)

check validity of specified valence space

NuclearToolkit.update_vsspace_chs! — Method

update_vsspace_chs!(HFobj,valencespace,Chan2b)

overwrite cc/vc/qc/vv/qv/qq channals

NuclearToolkit.write_vs_snt — Method

write_vs_snt(binfo,HFobj,IMSRGobj,Operators,effOps,Chan1b,Chan2bD,vspace)

Function to write out valence space effective interaction in snt (KSHELL/ShellModel.jl) format.